commit 2

+++ src/main/java/com/redvas/app/ui/UITool.java
@@ -4,10 +4,10 @@ import java.awt.image.BufferedImage;
 
 public class UITool {
     public static int fitWidth2AspectRatio(BufferedImage img, int height) {
-        return img.getWidth() / img.getHeight() * height;
+        return (int)((double)img.getWidth() / img.getHeight() * height);
     }
 
     public static int fitHeight2AspectRatio(BufferedImage img, int width) {
-        return img.getHeight() / img.getWidth() * width;
+        return (int)((double)img.getHeight() / img.getWidth() * width);
     }
 }
